Fix conversion of timeout to clock cycles
Currently the timeout provided to spinKernel is in nanoseconds.

There are 1e6 nanoseconds in a millisecond, but the implementation
divides timeoutNs by 1000 to convert to milliseconds as part of the
conversion from nanoseconds to clock cycles:

    timeoutClocks = (clocksPerMs * timeoutNs) / 1000

so the resulting timeout is 10000 seconds (~ 2.7 hours) instead of 10 seconds.

The direct fix is to correct the ns -> ms conversion factor:

    timeoutClocks = (clocksPerMs * timeoutNs) / 1000000

but this spinKernel timeout value doesn't use or need nanosecond resolution.

This change fixes the conversion of the timeout from time to clock
cycles by modifying spinKernel to take timeout in milliseconds
instead, which results in this conversion:

    timeoutClocks = clocksPerMs * timeoutMs